backend/requirements.txt
mcp[fastapi,server]
pydantic
langchain
langchain-core
langchain-google-genai
python-dotenv
fastapi
uvicorn
